chromium/third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/cross-origin-isolated-permission.https-expected.txt

This is a testharness.js-based test.
FAIL origin = https://web-platform.test:8444, value = undefined assert_equals: expected true but got false
FAIL origin = https://web-platform.test:8444, value = * assert_equals: expected true but got false
FAIL origin = https://web-platform.test:8444, value = 'self' assert_equals: expected true but got false
PASS origin = https://web-platform.test:8444, value = 'none'
PASS origin = https://www1.web-platform.test:8444, value = undefined
FAIL origin = https://www1.web-platform.test:8444, value = * assert_equals: expected true but got false
PASS origin = https://www1.web-platform.test:8444, value = 'self'
PASS origin = https://www1.web-platform.test:8444, value = 'none'
Harness: the test ran to completion.
